+++ tests/integration_tests/setup_window_tests.py
@@ -2,7 +2,6 @@ import os
 import time
 import unittest
 import warnings
-from shutil import which
 
 from dotenv import load_dotenv
 
@@ -22,7 +21,7 @@ def remove_db_file():
         os.remove(db_file_path)
 
 
-class MainTestCase(unittest.TestCase):
+class SetupWindowTestCase(unittest.TestCase):
     @classmethod
     def setUpClass(cls) -> None:
         remove_db_file()
@@ -56,21 +55,7 @@ class MainTestCase(unittest.TestCase):
             else:
                 break
 
-    def test_environ_vars_available(self) -> None:
-        """ Test if the environment variables a present. """
-        self.assertIsNotNone(os.environ.get('AWS_ACCESS_KEY_ID'))
-        self.assertIsNotNone(os.environ.get('AWS_SECRET_KEY'))
-        self.assertIsNotNone(os.environ.get('AWS_REGION_NAME'))
-
-    def test_ffmpeg_in_path(self) -> None:
-        """ Test if ffmpeg is added to path. """
-        self.assertIsNotNone(which('ffmpeg'))
-
-    def test_ffprobe_in_path(self) -> None:
-        """ Test if ffprobe is added to path. """
-        self.assertIsNotNone(which('ffprobe'))
-
-    def test_setup_window(self) -> None:
+    def test_setup_window_on_fresh_database(self) -> None:
         """ Test that everyting in the setup window is working correctly. """
 
         self.pc = ProgramController()
